pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "diffinst"
version = "0.0.1"
requires-python = ">=3.9"
dependencies = ["numpy", "pyyaml", "scipy", "matplotlib"]

[tool.setuptools.packages.find]
include = ["diffinst*"]